feat: test mock create
